example/serve-fastapi/pyproject.toml
[project]
name = "serve-fastapi"
version = "0.1.0"
description = "FastAPI service for serving ML models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "numpy>=1.26.0",
    "mlflow-skinny>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
